SPIP ne pose pas de problèmes avec des fichiers de plus de 2Go sauf que le champ "taille" de spip_documents est un int et que sa valeur maximale est limitée ... On passe donc en bigint

+++ base/medias.php
@@ -127,7 +127,7 @@ function medias_declarer_tables_objets_sql($tables){
 			"date"	=> "datetime DEFAULT '0000-00-00 00:00:00' NOT NULL",
 			"descriptif"	=> "text DEFAULT '' NOT NULL",
 			"fichier"	=> "text NOT NULL DEFAULT ''",
-			"taille"	=> "integer",
+			"taille"	=> "bigint",
 			"largeur"	=> "integer",
 			"hauteur"	=> "integer",
 			"mode"	=> "varchar(10) DEFAULT 'document' NOT NULL",
